refactor, test(xod-func-tools): move functions into separate modules, cover almost all functions (complex or derivative), move tests into separate specs, add `foldMaybe`
