+++ packages/xod-client/src/types.js
@@ -1,7 +1,7 @@
 import R from 'ramda';
 import $ from 'sanctuary-def';
 
-import XF from 'xod-func-tools';
+import * as XF from 'xod-func-tools';
 import * as XP from 'xod-project';
 
 import { SELECTION_ENTITY_TYPE } from './editor/constants';
