+++ packages/xod-fs/src/types.js
@@ -2,7 +2,7 @@ import R from 'ramda';
 import $ from 'sanctuary-def';
 import HMDef from 'hm-def';
 import { Identifier, Patch, Node, Link, Comment, env as xpEnv } from 'xod-project';
-import XF from 'xod-func-tools';
+import * as XF from 'xod-func-tools';
 
 /* Types are by convention starts with a capital leter, so: */
 /* eslint-disable new-cap */
