+++ packages/xod-project/src/types.js
@@ -2,7 +2,7 @@ import HMDef from 'hm-def';
 import R from 'ramda';
 import $ from 'sanctuary-def';
 
-import XF from 'xod-func-tools';
+import * as XF from 'xod-func-tools';
 
 import * as C from './constants';
 import { isTerminalPatchPath, isValidIdentifier, isValidPatchPath } from './internal/patchPathUtils';
